Headless Chrome initial implementation
This draft of the module includes some of the API functions, some are still pending.
Anything related to Exception Handling, dependency checks, etc, hasn't been implemented yet.
The class constructor calls the functions defined in Module.py, however the class itself doesn't inherit
from the Module class yet. Some Interactions (AKA actions) haven't been implemented, and the rest are just wrappers
around the ActionChains class. As this is a separate object, it may be moved to a different "Interactions" class to keep everything tidy.

The configuration uses a config.json file to get the driver and Chrome executables paths, as well as the working URL. Some options, like window_size, are going to be implemented later. It is possible multiple modules will share the same config, so the JSON file would include objects per module like so:

config.json:
```
{ HeadlessChrome: { "chromedriver_path":"...",
                     "url": "..." }
  OtherModule: ... }
```
It is possible that a JSON file might not be used at all for modules in the production version of Pyroute.

This code has been linted using Flake8, however some style issues may be still present, like name conventions.