200_Colors Added custom colors to the terminal.
The json configuration file has a new key named color, in this key you can
choose what color you want for each thing or if you don't want color at all.

This feature will be activated if the "color" property is "on", anything else
will disable it.
I uploaded a configuration file to the wiki so it works the same for everyone,
the colors on this file are just defaults, you can change it however you want.

Here's a list of all the colors and styles you can use as of now:
"BRG" stands for bright color
"BG" stands for the background color

 -BLACK
 -RED
 -GREEN
 -ORANGE
 -BLUE
 -MAGENTA
 -CYAN
 -GREY
 -BRG_BLACK
 -BRG_RED
 -BRG_GREEN
 -BRG_YELLOW
 -BRG_BLUE
 -BRG_MAGENTA
 -BRG_CYAN
 -BRG_WHITE
 -BG_BLACK
 -BG_RED
 -BG_GREEN
 -BG_YELLOW
 -BG_BLUE
 -BG_MAGENTA
 -BG_CYAN
 -BG_WHITE
 -BG_BRG_BLACK
 -BG_BRG_RED
 -BG_BRG_GREEN
 -BG_BRG_YELLOW
 -BG_BRG_BLUE
 -BG_BRG_MAGENTA
 -BG_BRG_CYAN
 -BG_BRG_GREY
 -BOLD
 -UNDERLINE
 -RESET